backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
ultralytics==8.0.20
timm==0.9.2
torch==2.0.1
torchvision==0.15.2
opencv-python==4.8.0.74
pillow==9.5.0
numpy==1.24.3
python-dotenv==1.0.0
gunicorn==21.2.0
